requirements.txt
asgiref==3.3.4
attrs==20.3.0
Baker==1.3
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
drf-writable-nested==0.6.2
idna==2.10
inflection==0.3.1
iniconfig==1.1.1
lazy-object-proxy==1.6.0
model-bakery==1.3.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytest-assert-utils==0.2.2
pytest-common-subject==1.0.5
pytest-django==4.2.0
pytest-fixture-order==0.1.3
pytest-lambda==1.2.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
